fix: initialize ServiceLoaders from the FunctionFactory class loader (#1172)
In environments where the QuestDb classes are not loaded by the system
class loader and no context class loader is explicitly set, functions
will not be found when initializing a ServiceLoader without passing
a class loader explicitly.

This change will always work assuming that SQL functions are loadable
from the same class loader as FunctionFactory, that is, there is no support
for plugging in external functions by passing a class loader capable of loading
those classes. I think that assumption is true for QuestDb as I can't find
any mention of class loaders in the QuestDb doc, and it would be challenging to
make this work because the external class loader would also need to correctly
load the functions bundled with QuestDb.